RSC convert admin services page
